requirements.txt
mlflow==2.10.2
scikit-learn
boto3
pandas
numpy
python-dotenv
setuptools